덧글 및 기록 수정/삭제 예외처리

+++ src/main/java/com/example/studymate/Study/service/CommentService.java
@@ -8,6 +8,7 @@ import com.example.studymate.Study.repository.CommentRepository;
 import com.example.studymate.User.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -45,11 +46,16 @@ public class CommentService {
         return CommentResponse.from(comment);
     }
 
+    @Transactional
     public CommentResponse updateComment(Long commentId,
                                          UpdateCommentRequest request,
                                          User user) {
         StudyComment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 id의 덧글이 없습니다."));
+
+        if (!comment.getAuthorId().equals(user.getId())) {
+            throw new IllegalArgumentException("덧글 수정 권한이 없습니다.");
+        }
 
         comment.update(request.getContent());
         return CommentResponse.from(comment);
@@ -59,6 +65,10 @@ public class CommentService {
                               User user) {
         StudyComment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 id의 덧글이 없습니다."));
+
+        if (!comment.getAuthorId().equals(user.getId())) {
+            throw new IllegalArgumentException("덧글 삭제 권한이 없습니다.");
+        }
 
         commentRepository.deleteById(commentId);
     }
